Multi-Geo
I am requesting we review that unified audit logs at this time are not working in a multi-Geo tenant for exchange admin audits. I am also requesting we get public documentation published to let our customers know this is not working for admins. 

There is another article that states clearly Cross-Geo mailbox auditing is not supported. So there is a ton of confusion here that we must clean up